+++ src/Set/Drupal12SetList.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DrupalRector\Set;
+
+final class Drupal12SetList
+{
+    public const DRUPAL_12 = __DIR__.'/../../config/drupal-12/drupal-12-all-deprecations.php';
+    public const DRUPAL_120 = __DIR__.'/../../config/drupal-12/drupal-12.0-deprecations.php';
+}